+++ minifica.api/Controllers/SchoolBillingInformationController.cs
@@ -23,7 +23,7 @@ namespace minifica.api.Controllers
         [HttpPatch("{id}")]
         public override IActionResult Update(int id, SchoolBillingInformationViewModel schoolBillingInformation) => base.Update(id, schoolBillingInformation);
 
-        [HttpDelete]
+        [HttpDelete("{id}")]
         public override IActionResult Delete(int id) => base.Delete(id);
     }
 }